requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.3
django-bootstrap-form==3.4
django-discover-runner==1.0
django-heroku==0.3.1
django-localflavor==3.1
django-role-permissions==3.1.1
django-widget-tweaks==1.4.12
fontawesomefree==6.2.0
gunicorn==20.1.0
mysqlclient==2.1.1
Pillow==9.3.0
psycopg2-binary==2.9.5
pypi==2.1
python-dateutil==2.8.2
python-decouple==3.6
python-stdnum==1.17
six==1.16.0
sqlparse==0.4.3
whitenoise==6.2.0
